added lights
